pyproject.toml
[project]
name = "predictions"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
